actions: Split out zerver.actions.video_calls.
