+++ zerver/views/video_calls.py
@@ -22,8 +22,8 @@ from django.views.decorators.http import require_POST
 from oauthlib.oauth2 import OAuth2Error
 from requests_oauthlib import OAuth2Session
 
+from zerver.actions.video_calls import do_set_zoom_token
 from zerver.decorator import zulip_login_required
-from zerver.lib.actions import do_set_zoom_token
 from zerver.lib.exceptions import ErrorCode, JsonableError
 from zerver.lib.outgoing_http import OutgoingSession
 from zerver.lib.pysa import mark_sanitized
